+++ src/adjusters/boolean.mjs
@@ -1,9 +1,9 @@
 import AdjusterBase from "../libs/AdjusterBase";
 
-import Default from "../libs/decorators/default";
-import AcceptNull from "../libs/decorators/acceptNull";
-import AcceptEmptyString from "../libs/decorators/acceptEmptyString";
-import Type from "../libs/decorators/boolean/type";
+import Default from "../decorators/default";
+import AcceptNull from "../decorators/acceptNull";
+import AcceptEmptyString from "../decorators/acceptEmptyString";
+import Type from "../decorators/boolean/type";
 
 /**
  * factory
