+++ src/adjusters/ipv4.mjs
@@ -1,11 +1,11 @@
 import AdjusterBase from "../libs/AdjusterBase";
 
-import Default from "../libs/decorators/default";
-import AcceptNull from "../libs/decorators/acceptNull";
-import AcceptEmptyString from "../libs/decorators/acceptEmptyString";
-import Type from "../libs/decorators/string/type";
-import Trim from "../libs/decorators/string/trim";
-import Pattern from "../libs/decorators/string/pattern";
+import Default from "../decorators/default";
+import AcceptNull from "../decorators/acceptNull";
+import AcceptEmptyString from "../decorators/acceptEmptyString";
+import Type from "../decorators/string/type";
+import Trim from "../decorators/string/trim";
+import Pattern from "../decorators/string/pattern";
 
 const PATTERN_COMPONENT = `(25[0-5]|2[0-4][0-9]|[0-1]?[0-9]{1,2})`;
 const PATTERN = `${PATTERN_COMPONENT}(\\.${PATTERN_COMPONENT}){3}`;
