+++ src/adjusters/number.mjs
@@ -1,12 +1,12 @@
 import AdjusterBase from "../libs/AdjusterBase";
 
-import Default from "../libs/decorators/default";
-import AcceptNull from "../libs/decorators/acceptNull";
-import AcceptEmptyString from "../libs/decorators/acceptEmptyString";
-import Only from "../libs/decorators/only";
-import Type from "../libs/decorators/number/type";
-import MinValue from "../libs/decorators/number/minValue";
-import MaxValue from "../libs/decorators/number/maxValue";
+import Default from "../decorators/default";
+import AcceptNull from "../decorators/acceptNull";
+import AcceptEmptyString from "../decorators/acceptEmptyString";
+import Only from "../decorators/only";
+import Type from "../decorators/number/type";
+import MinValue from "../decorators/number/minValue";
+import MaxValue from "../decorators/number/maxValue";
 
 /**
  * factory
