+++ src/adjusters/object.mjs
@@ -1,10 +1,10 @@
 import AdjusterBase from "../libs/AdjusterBase";
 
-import Default from "../libs/decorators/default";
-import AcceptNull from "../libs/decorators/acceptNull";
-import AcceptEmptyString from "../libs/decorators/acceptEmptyString";
-import Type from "../libs/decorators/object/type";
-import Constraints from "../libs/decorators/object/constraints";
+import Default from "../decorators/default";
+import AcceptNull from "../decorators/acceptNull";
+import AcceptEmptyString from "../decorators/acceptEmptyString";
+import Type from "../decorators/object/type";
+import Constraints from "../decorators/object/constraints";
 
 /**
  * factory
