+++ src/adjusters/string.mjs
@@ -1,14 +1,14 @@
 import AdjusterBase from "../libs/AdjusterBase";
 
-import Default from "../libs/decorators/default";
-import AcceptNull from "../libs/decorators/acceptNull";
-import AcceptEmptyString from "../libs/decorators/acceptEmptyString";
-import Only from "../libs/decorators/only";
-import Type from "../libs/decorators/string/type";
-import Trim from "../libs/decorators/string/trim";
-import MinLength from "../libs/decorators/string/minLength";
-import MaxLength from "../libs/decorators/string/maxLength";
-import Pattern from "../libs/decorators/string/pattern";
+import Default from "../decorators/default";
+import AcceptNull from "../decorators/acceptNull";
+import AcceptEmptyString from "../decorators/acceptEmptyString";
+import Only from "../decorators/only";
+import Type from "../decorators/string/type";
+import Trim from "../decorators/string/trim";
+import MinLength from "../decorators/string/minLength";
+import MaxLength from "../decorators/string/maxLength";
+import Pattern from "../decorators/string/pattern";
 
 /**
  * factory
