+++ src/decorators/default.mjs
@@ -1,0 +1,56 @@
+import {CAUSE} from "../libs/constants";
+import AdjusterBase from "../libs/AdjusterBase";
+import AdjusterError from "../libs/AdjusterError";
+
+export default AdjusterBase.decoratorBuilder(_adjust)
+	.init(_init)
+	.features({
+		default: _featureDefault,
+	})
+	.build();
+
+/**
+ * init
+ * @param {Object} params parameters
+ * @returns {void}
+ */
+function _init(params)
+{
+	params.flag = false;
+}
+
+/**
+ * set default value
+ * @param {Object} params parameters
+ * @param {*} value default value
+ * @returns {void}
+ */
+function _featureDefault(params, value)
+{
+	params.flag = true;
+	params.value = value;
+}
+
+/**
+ * adjuster
+ * @param {Object} params parameters
+ * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
+ * @returns {boolean} end adjustment
+ * @throws {AdjusterError}
+ */
+function _adjust(params, values, keyStack)
+{
+	if(values.adjusted !== undefined)
+	{
+		return false;
+	}
+
+	if(params.flag)
+	{
+		values.adjusted = params.value;
+		return true;
+	}
+
+	AdjusterError.raise(CAUSE.REQUIRED, values, keyStack);
+}
