+++ src/decorators/number/minValue.mjs
@@ -1,0 +1,56 @@
+import {CAUSE} from "../../libs/constants";
+import AdjusterBase from "../../libs/AdjusterBase";
+import AdjusterError from "../../libs/AdjusterError";
+
+export default AdjusterBase.decoratorBuilder(_adjust)
+	.init(_init)
+	.features({
+		minValue: _featureMinValue,
+	})
+	.build();
+
+/**
+ * init
+ * @param {Object} params parameters
+ * @returns {void}
+ */
+function _init(params)
+{
+	params.value = Number.MIN_SAFE_INTEGER;
+}
+
+/**
+ * set min-value
+ * @param {Object} params parameters
+ * @param {number} value min-value
+ * @param {boolean} [adjust=false] adjust to min-value if value < min-value; default is ERROR
+ * @returns {void}
+ */
+function _featureMinValue(params, value, adjust = false)
+{
+	params.value = value;
+	params.adjust = adjust;
+}
+
+/**
+ * adjust
+ * @param {Object} params parameters
+ * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
+ * @returns {boolean} end adjustment
+ * @throws {AdjusterError}
+ */
+function _adjust(params, values, keyStack)
+{
+	if(values.adjusted >= params.value)
+	{
+		return false;
+	}
+	if(params.adjust)
+	{
+		values.adjusted = params.value;
+		return false;
+	}
+
+	AdjusterError.raise(CAUSE.MIN_VALUE, values, keyStack);
+}
