--- src/libs/decorators/array/minLength.mjs
@@ -1,54 +1,0 @@
-import {CAUSE} from "../../constants";
-import AdjusterBase from "../../AdjusterBase";
-import AdjusterError from "../../AdjusterError";
-
-export default AdjusterBase.decoratorBuilder(_adjust)
-	.init(_init)
-	.features({
-		minLength: _featureMinLength,
-	})
-	.build();
-
-/**
- * init
- * @param {Object} params parameters
- * @returns {void}
- */
-function _init(params)
-{
-	params.flag = false;
-}
-
-/**
- * set min-length of array
- * @param {Object} params parameters
- * @param {number} length max-length
- * @returns {void}
- */
-function _featureMinLength(params, length)
-{
-	params.flag = true;
-	params.length = length;
-}
-
-/**
- * adjuster
- * @param {Object} params parameters
- * @param {AdjusterBase.VALUES} values original / adjusted values
- * @param {(string|number)[]} keyStack path to key that caused error
- * @returns {boolean} end adjustment
- * @throws {AdjusterError}
- */
-function _adjust(params, values, keyStack)
-{
-	if(!params.flag)
-	{
-		return false;
-	}
-	if(values.adjusted.length >= params.length)
-	{
-		return false;
-	}
-
-	AdjusterError.raise(CAUSE.MIN_LENGTH, values, keyStack);
-}
