--- src/libs/decorators/object/type.mjs
@@ -1,25 +1,0 @@
-import {CAUSE} from "../../constants";
-import {isObject} from "../../types";
-import AdjusterBase from "../../AdjusterBase";
-import AdjusterError from "../../AdjusterError";
-
-export default AdjusterBase.decoratorBuilder(_adjust)
-	.build();
-
-/**
- * adjuster
- * @param {Object} params parameters
- * @param {AdjusterBase.VALUES} values original / adjusted values
- * @param {(string|number)[]} keyStack path to key that caused error
- * @returns {boolean} end adjustment
- * @throws {AdjusterError}
- */
-function _adjust(params, values, keyStack)
-{
-	if(isObject(values.adjusted))
-	{
-		return false;
-	}
-
-	AdjusterError.raise(CAUSE.TYPE, values.original, keyStack);
-}
